Structures for nonblocking IO.
Each filehandle is associated with a ksudo_fd structure, which holds
buffers and a vtable, and a struct pollfd to pass to poll. The pollfds
are held in a separate array so that they can be passed to poll(2); this
means that functions and macros operating on a ksfd must identify it by
index rather than using a pointer.

The buffers are optional since listen fds will never need to do ordinary
IO.