Forget the ring buffers.
Since the krb5 functions all want their buffers to be contiguous anyway,
we might as well keep them that way from the beginning. Instead of a
ring buffer, use a buffer with a start and end pointer, plus a macro to
shift the valid part of the buffer back to the start if we need to.
Currently I keep the buffer twice the size it strictly needs to be to
avoid having to use memmove(3); this might be silly.