Buffers for writing ASN.1.
The ASN.1 Kerberos structures need different buffer arrangements from
ordinary data, at least for writing, since the Krb functions insist on
allocating their buffers themselves. A ksudod_msgbuf holds two pointer
to krb5_data structures, the one we are currently writing and the next
one to be written, and a pointer into the first which points at the next
byte to write. (The reason for having two is so that we don't have to
make a write of just a few bytes when we are near the end of a
structure.)

There are no actual IO functions yet, pending the ksfd reorganisation.