Rework the ksudo_fd structure.
The read and write buffers are no longer a fixed part of the structure;
instead, they are part of the type-specific data structure. This
requires additional 'open' and 'close' vtable entries, to handle
initialising and freeing these structures.

I've brought back the concept of 'type', since callers need to know
they are writing to the right type of ksfd, but there's no need for a
separate field since I can just check the address assigned to the 'ops'
field.

This also implements the fdops for KSFt(msg), though since I don't have
a 'session' structure yet the read op doesn't do anything with the data.