Set and clear the event bits in the poll structs.
Before we ended up busy-waiting around a poll that kept saying 'yes,
this filehandle is writable' when we didn't have anything to write.

The interface only allows setting/clearing/questioning one direction at
a time, since I suspect I'm going to have to port all this over to
pselect(2) at some point to get sane signal handling. (The other option
would be to use a self-pipe, which may be better.)