Generic signal handling.
The client and server should not share the SIGCHLD handler, since
otherwise the client might start sending KSUDO-EXITs to the server,
which would be very wrong. In any case, both client and server will need
a general signal-handling mechanism, since the client need to send
KSUDO-SIGNALs and the server needs to handle HUP, TERM, &c. as a daemon
should.

Implemented like this, using volatile globals and poll(2), there is a
race which means poll may not end up interrupted. I need to fix this by
either switching to pselect(2) or using the self-pipe trick.