Map signals to KSUDO-SIGNALs in kss_exit.
I've also removed the nonfatal signals from KSUDO-SIGNAL, since they
shouldn't appear as the WTERMSIG in an exit status. This may turn out to
be a mistake later when I start mapping signal reception from client to
server, but I'm not entirely sure how to make that work (for instance,
if the client process is stopped it can't receive a SIGCONT over the
wire...). I've also added SIGSYS, which is in SUSv4, and potentially
important.